+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add `PUT /accounts/:id` endpoint
 - Fixed bug where child account's DCEPrincipal role trusted itself rather than the master account
 - Add GetUsageByPrincipal
+- Fix default `budget_notification_from_email` TF var (See #143)
 
 
 ## v0.23.0